requirements.txt
h11==0.16.0
pip==23.2.1
attrs==25.3.0
wheel==0.41.2
PySocks==1.7.1
cffi==1.17.1
setuptools==68.2.0
typing_extensions==4.13.2
pytest==8.3.4
trio==0.30.0
outcome==1.3.0.post0
sniffio==1.3.1
exceptiongroup==1.2.2
sortedcontainers==2.4.0
idna==3.10
mysql-connector-python==9.3.0
allure-python-commons==2.14.2
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3==1.26.18
pluggy==1.5.0
iniconfig==2.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
packaging==24.2
pycparser==2.22
websocket-client==1.8.0
eyes-selenium==6.2.0
marshmallow==3.26.1
allure-pytest==2.14.2
wsproto==1.2.0
requests==2.32.3
webdriver-manager==4.0.2
python-dotenv==1.0.1